Add selective CRUD

+++ src/pets/pets.controller.ts
@@ -6,14 +6,19 @@ import {
   Patch,
   Param,
   Delete,
+  ForbiddenException,
 } from '@nestjs/common';
 import { PetsService } from './pets.service';
 import { CreatePetDto } from './dto/create-pet.dto';
 import { UpdatePetDto } from './dto/update-pet.dto';
+import { UsersService } from 'src/users/users.service';
 
 @Controller('pets')
 export class PetsController {
-  constructor(private readonly petsService: PetsService) {}
+  constructor(
+    private readonly petsService: PetsService,
+    private readonly users: UsersService,
+  ) {}
 
   @Post()
   create(@Body() createPetDto: CreatePetDto) {
@@ -31,12 +36,31 @@ export class PetsController {
   }
 
   @Patch(':id')
-  update(@Param('id') id: string, @Body() updatePetDto: UpdatePetDto) {
-    return this.petsService.update(id, updatePetDto);
+  async update(@Param('id') id: string, @Body() data: UpdatePetDto) {
+    const userEmail = data.email;
+    if (!userEmail || !this.users.isUserLogged(userEmail)) {
+      throw new ForbiddenException('Access Denied');
+    }
+    const pet = this.petsService.findOne(id);
+    if ((await pet).userId !== userEmail) {
+      throw new ForbiddenException('You can only update your pets');
+    }
+    delete data.email;
+    return this.petsService.update(id, data);
   }
 
   @Delete(':id')
-  remove(@Param('id') id: string) {
+  async remove(
+    @Param('id') id: string,
+    @Body() { userEmail }: { userEmail: string },
+  ) {
+    if (!this.users.isUserLogged(userEmail)) {
+      throw new ForbiddenException('Access Denied');
+    }
+    const pet = this.petsService.findOne(id);
+    if ((await pet).userId !== userEmail) {
+      throw new ForbiddenException('You can only remove your pets');
+    }
     return this.petsService.remove(id);
   }
 }
